feat: refactor user schemas and controller for improved type safety and rename response helper utility
